+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 The format of this file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0] - 2019-04-30
+
+### Added
+
+- New API function `preloadMultiple` for loading multiple sounds in one shot. A good
+  alternative to multiple `preloadSound` calls if you want to load many assets and be
+  notified once, when they are all complete. The trade-off is less detailed error-handling.
+
 ## [1.1.0] - 2019-04-29
 
 ### Added
@@ -35,6 +43,7 @@ The format of this file is based on [Keep a Changelog](https://keepachangelog.co
 - First functional release of cordova-plugin-ui-sounds! See the [README](README.md)
   for full details.
 
-[unreleased]: https://github.com/talisman-games/cordova-plugin-ui-sounds/compare/1.1.0...HEAD
+[unreleased]: https://github.com/talisman-games/cordova-plugin-ui-sounds/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/talisman-games/cordova-plugin-ui-sounds/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/talisman-games/cordova-plugin-ui-sounds/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/talisman-games/cordova-plugin-ui-sounds/releases/tag/1.0.0